Add service.app_controller module
