+++ .mypy.ini
@@ -51,7 +51,8 @@ modules =
     azul.indexer.document,
     azul.indexer.field,
     azul.indexer.action_controller,
-    azul.indexer.mirror_controller
+    azul.indexer.mirror_controller,
+    azul.service.app_controller
 packages =
     azul.openapi
